+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.5] - 2024-10-14
+
+This update introduces a critical bug fix
+
+
+
 ## [2.1.4] - 2024-10-13
 
 This update introduces minor bug fixes & improvements